+++ src/main/java/com/backfcdev/managementsystem/repository/ICategoryRepository.java
@@ -2,5 +2,5 @@ package com.backfcdev.managementsystem.repository;
 
 import com.backfcdev.managementsystem.model.Category;
 
-public interface ICategoryRepository extends IGenericRepository<Category, Integer> {
+public interface ICategoryRepository extends IGenericRepository<Category, Long> {
 }
