+++ src/main/java/com/backfcdev/managementsystem/repository/IOrderRepository.java
@@ -2,5 +2,5 @@ package com.backfcdev.managementsystem.repository;
 
 import com.backfcdev.managementsystem.model.Order;
 
-public interface IOrderRepository extends IGenericRepository<Order, Integer>{
+public interface IOrderRepository extends IGenericRepository<Order, Long>{
 }
