+++ src/main/java/com/backfcdev/managementsystem/repository/IProductRepository.java
@@ -4,6 +4,6 @@ import com.backfcdev.managementsystem.model.Product;
 
 import java.util.List;
 
-public interface IProductRepository extends IGenericRepository<Product, Integer> {
+public interface IProductRepository extends IGenericRepository<Product, Long> {
     List<Product> findByStockLessThan(int amount);
 }
